# Improve cache locality and perf of DeepGru on CPU